lib: converge.rs: implement find_dominator_value
The main functions are:
* find_divergent_changes
* choose_change
* converge_change
* apply_solution

This is going to be used to implement a jj converge command. See https://docs.jj-vcs.dev/latest/design/jj-converge-command/
Some functionality will be implemented in later commits.

Converge values using the value history graph: If T is the type of the value (String for description, Signature for author, Vec<CommitId> for parents) a Merge<T> is created using the closest common dominator of the values of the divergent commits over a "value history graph".